+++ src/main/kotlin/com/cycode/plugin/components/toolWindow/components/violationCardContentTab/sastViolationCardContentTab/components/companyGuidelines/SastCompanyGuidelines.kt
@@ -1,18 +1,18 @@
-package com.cycode.plugin.components.toolWindow.components.violationCardContentTab.iacViolationCardContentTab.components.companyGuidelines
+package com.cycode.plugin.components.toolWindow.components.violationCardContentTab.sastViolationCardContentTab.components.companyGuidelines
 
 import com.cycode.plugin.CycodeBundle
-import com.cycode.plugin.cli.models.scanResult.iac.IacDetection
+import com.cycode.plugin.cli.models.scanResult.sast.SastDetection
 import com.cycode.plugin.components.toolWindow.components.violationCardContentTab.common.htmlSummary.CardHtmlSummary
 import com.cycode.plugin.components.toolWindow.components.violationCardContentTab.convertMarkdownToHtml
 import javax.swing.JComponent
 
-class IacCompanyGuidelines : CardHtmlSummary() {
-    private fun getCustomGuidelines(detection: IacDetection): String? {
+class SastCompanyGuidelines : CardHtmlSummary() {
+    private fun getCustomGuidelines(detection: SastDetection): String? {
         val descriptionMarkdown = detection.detectionDetails.customRemediationGuidelines ?: return null
         return convertMarkdownToHtml(descriptionMarkdown)
     }
 
-    fun getContent(detection: IacDetection): JComponent {
+    fun getContent(detection: SastDetection): JComponent {
         return getContent(
             CycodeBundle.message("violationCardCompanyGuidelinesTitle"),
             getCustomGuidelines(detection)