+++ src/main/kotlin/com/cycode/plugin/components/toolWindow/components/violationCardContentTab/secretViolationCardContentTab/components/cycodeGuidelines/SecretCycodeGuidelines.kt
@@ -1,18 +1,18 @@
-package com.cycode.plugin.components.toolWindow.components.violationCardContentTab.iacViolationCardContentTab.components.cycodeGuidelines
+package com.cycode.plugin.components.toolWindow.components.violationCardContentTab.secretViolationCardContentTab.components.cycodeGuidelines
 
 import com.cycode.plugin.CycodeBundle
-import com.cycode.plugin.cli.models.scanResult.iac.IacDetection
+import com.cycode.plugin.cli.models.scanResult.secret.SecretDetection
 import com.cycode.plugin.components.toolWindow.components.violationCardContentTab.common.htmlSummary.CardHtmlSummary
 import com.cycode.plugin.components.toolWindow.components.violationCardContentTab.convertMarkdownToHtml
 import javax.swing.JComponent
 
-class IacCycodeGuidelines : CardHtmlSummary() {
-    private fun getCycodeGuidelines(detection: IacDetection): String? {
+class SecretCycodeGuidelines : CardHtmlSummary() {
+    private fun getCycodeGuidelines(detection: SecretDetection): String? {
         val descriptionMarkdown = detection.detectionDetails.remediationGuidelines ?: return null
         return convertMarkdownToHtml(descriptionMarkdown)
     }
 
-    fun getContent(detection: IacDetection): JComponent {
+    fun getContent(detection: SecretDetection): JComponent {
         return getContent(
             CycodeBundle.message("violationCardCycodeGuidelinesTitle"),
             getCycodeGuidelines(detection)